[account] extract txConfirmLogic
